chrome-backend/requirements.txt
blinker==1.6.2
boto3==1.28.57
botocore==1.31.57
click==8.1.7
flask==2.2.5
Flask-Cors==3.0.10
gunicorn==21.2.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
s3transfer==0.7.0
six==1.16.0
urllib3==1.26.16
werkzeug==2.2.3
zipp==3.15.0
